+++ src/utils/Math.sol
@@ -1,11 +1,11 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.24;
 
-function add(uint96 a, int192 b) pure returns (uint96) {
+function add(uint88 a, int192 b) pure returns (uint88) {
     if (b < 0) {
-        return uint96(a - uint96(uint192(-b)));
+        return uint88(a - uint88(uint192(-b)));
     }
-    return uint96(a + uint96(uint192(b)));
+    return uint88(a + uint88(uint192(b)));
 }
 
 function sub(uint256 a, int256 b) pure returns (uint128) {
